+++ worker/wrangler.toml.template
@@ -1,6 +1,6 @@
 name = "cloudflare_temp_email"
 main = "src/worker.js"
-compatibility_date = "2023-08-14"
+compatibility_date = "2023-12-01"
 node_compat = true
 
 [vars]
